pyproject.toml
[project]
name = "synthea-hospital-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "pydantic>=2.12.3",
    "pyspark>=4.0.1",
    "python-dotenv>=1.1.1",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "supabase>=2.28.3",
]
